Check for jammed/frozen before each attack in attack_phase
This accounts for a few problems:
If a Chaosed assault uses Freeze or Jam on itself, it should not be able
to attack, but it was previously able to.
If an assault attacks, triggering a Jam on Attacked, it should not be
able to attack, but it was previously able to.